Refactor cipher files
