+++ infrastructure/tari_util/src/ciphers/mod.rs
@@ -18,20 +18,7 @@
 // SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
 // SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
-// USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-#[allow(clippy::needless_range_loop)]
-pub mod bit;
-pub mod byte_array;
-pub mod ciphers;
-pub mod extend_bytes;
-pub mod fixed_set;
-pub mod hash;
-pub mod hex;
-pub mod message_format;
+// USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE
 
-pub use self::extend_bytes::ExtendBytes;
-
-pub use self::{
-    byte_array::{ByteArray, ByteArrayError},
-    hash::Hashable,
-};
+pub mod chacha20;
+pub mod cipher;
